ti33x: move screen feature to machine configs
The AM335x SoC has a built-in 16bit/24bit RGB LCD controller, and
the BeagleBone Black uses an external chip to convert this to HDMI.

But the SoC does not have a video output port per-se, thus the "screen"
machine feature does not belong to its configuration but rather to
the machine configurations.

Move the feature (and the MACHINE_GUI_CLASS definition) out of
ti33x.inc into am335x-evm.conf and beaglebone.conf respectively.

At the same time, drop the irrelevant comments regarding the external
screen, since MACHINE_GUI_CLASS is unlikely used and the BeagleBone
Black has a micro HDMI port, not a DVI port.

Signed-off-by: Vivien Didelot <vdidelot@pbsc.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>